pyproject.toml
[tool.poetry]
name = "movies-api"
version = "0.1.0"
description = "Movies API using FASTAPI as backend framework"
authors = ["cristian-rincon <cristian.o.rincon.b@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
loguru = "^0.6.0"
pyjwt = "^2.6.0"
sqlalchemy = "^1.4.46"


[tool.poetry.group.dev.dependencies]
black = {version = "^23.1a1", allow-prereleases = true}
isort = "^5.11.4"
pre-commit = "^2.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
